New structure, Mifiel module should include all classes
