--- lib/crypto.rb
@@ -1,5 +1,0 @@
-module Crypto
-  autoload :PBE, 'crypto/pbe'
-  autoload :Response, 'crypto/response'
-  autoload :AES, 'crypto/aes'
-end
